// src/app/group-admin/models/post-recommendation.model.ts
import {PostContentTypeEnum} from './facebook-post.model';

export class PostRecommendationModel {
	id: string;
	timings: string;
	toBePostedAtUTCTicks: number;
	contentType: PostContentTypeEnum;

	optLength: {
		minChar: number;
		maxChar: number;
		minWord: number;
		maxWord: number;
	};

	emotions: any;
	keywords: string[];
	categories: string[];
	topics: string[];
	iconUrl: string;
	headline: string;
	description: string;
	imageUrls: string[];

	type: RecommendationType;

	createdByUserId: string;
	createdAtUTC: string;

	isDeleted: boolean;

	// The below properties for the UI and not for backend
	toBePostedAtInLocalTime: string;
	source: RecommendationSourceEnum;
	toBePostedAtUTCHour: number;

	public makeItReadyForSendingToBackend() {
		delete this.source;
		delete this.createdAtUTC;
		delete this.toBePostedAtUTCHour;
		delete this.toBePostedAtInLocalTime;
	}
}

export enum RecommendationType {
	Timing = 'Timing',
	ContentType = 'ContentType'
}

export enum RecommendationSourceEnum {
	BDAdmin = 'BDAdmin',
	AutoGenerated = 'AutoGenerated'
}